+++ src/Pim/Bundle/EnrichBundle/Form/Subscriber/AddProductValueFieldSubscriber.php
@@ -2,13 +2,11 @@
 
 namespace Pim\Bundle\EnrichBundle\Form\Subscriber;
 
-use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
-use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Pim\Bundle\CatalogBundle\AttributeType\AttributeTypeFactory;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
+use Pim\Bundle\EnrichBundle\Form\Factory\ProductValueFormFactory;
 
 /**
  * Add a relevant form for each product value
@@ -20,25 +18,18 @@ use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 class AddProductValueFieldSubscriber implements EventSubscriberInterface
 {
     /**
-     * @var FormFactoryInterface
+     * @var ProductValueFormFactory
      */
     protected $factory;
 
     /**
-     * @var AttributeTypeFactory
-     */
-    protected $attributeTypeFactory;
-
-    /**
      * Constructor
      *
-     * @param FormFactoryInterface $factory
-     * @param AttributeTypeFactory $attTypeFactory
+     * @param ProductValueFormFactory $factory
      */
-    public function __construct(FormFactoryInterface $factory, AttributeTypeFactory $attTypeFactory)
+    public function __construct(ProductValueFormFactory $factory)
     {
         $this->factory = $factory;
-        $this->attributeTypeFactory = $attTypeFactory;
     }
 
     /**
@@ -68,11 +59,8 @@ class AddProductValueFieldSubscriber implements EventSubscriberInterface
             return;
         }
 
-        $attributeTypeAlias = $value->getAttribute()->getAttributeType();
-        $attributeType = $this->attributeTypeFactory->get($attributeTypeAlias);
+        $valueForm = $this->factory->buildProductValueForm($value);
 
-        /** @var FormInterface $valueForm */
-        $valueForm = $attributeType->buildValueFormType($this->factory, $value);
         $form->add($valueForm);
     }
 }