+++ CHANGELOG.md
@@ -27,6 +27,9 @@
 - Changed the HydratorInterface::hydrate() method signature
 - Avoid to store null values in Product::normalizedData (MongoDB support)
 - Remove redundant 'getActiveCodeChoices' method in CurrencyManager (use CurrencyManager::getActiveCodes())
+- Remove AbstractAttributeType::buildValueFormType, change visibility of prepareValueFormName, prepareValueFormAlias, prepareValueFormOptions, prepareValueFormConstraints, prepareValueFormData to public
+- Remove `MetricBaseValuesSubscriber` and create one for MongoDB and another one for ORM
+- Create `OptionFilter`, `OptionsFilter` for ORM and MongoDB implementations
 
 # 1.1.0 - "Rabbit Punch" (2014-04-16)